class

+++ 23.11/pjt/backend/articles/views.py
@@ -2,10 +2,6 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework import status
 
-# permission Decorators
-from rest_framework.decorators import permission_classes
-# from rest_framework.permissions import IsAuthenticated
-
 from django.shortcuts import get_object_or_404, get_list_or_404
 
 from .serializers import ArticleListSerializer, ArticleSerializer, CommentSerializer
@@ -13,7 +9,6 @@ from .models import Article, Comment
 
 
 @api_view(['GET', 'POST'])
-# @permission_classes([IsAuthenticated])
 def article_list(request):
     
     if request.method == 'GET':
@@ -30,7 +25,6 @@ def article_list(request):
 
 @api_view(['GET', 'DELETE', 'PUT'])
 def article_detail(request, article_pk):
-    # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
 
     if request.method == 'GET':
@@ -38,19 +32,22 @@ def article_detail(request, article_pk):
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
+        if request.user != article.user:
+            return Response({'detail': 'Permission denied.'}, status=status.HTTP_403_FORBIDDEN)
         article.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
+        if request.user != article.user:
+            return Response({'detail': 'Permission denied.'}, status=status.HTTP_403_FORBIDDEN)
         serializer = ArticleSerializer(article, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
-        # return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 @api_view(['GET'])
 def comment_list(request):
-    # comments = Comment.objects.all()
     comments = get_list_or_404(Comment)
     serializer = CommentSerializer(comments, many=True)
     return Response(serializer.data)
@@ -58,18 +55,22 @@ def comment_list(request):
 
 @api_view(['GET', 'DELETE', 'PUT'])
 def comment_detail(request, comment_pk):
-    # comment = Comment.objects.get(pk=comment_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
+
     if request.method == 'GET':
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
     
     elif request.method == 'DELETE':
+        if request.user != comment.user:
+            return Response({'detail': 'Permission denied.'}, status=status.HTTP_403_FORBIDDEN)
         comment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        serializer = CommentSerializer(comment, data=request.data)
+        if request.user != comment.user:
+            return Response({'detail': 'Permission denied.'}, status=status.HTTP_403_FORBIDDEN)
+        serializer = CommentSerializer(comment, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
@@ -77,9 +78,15 @@ def comment_detail(request, comment_pk):
 
 @api_view(['POST'])
 def comment_create(request, article_pk):
-    # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
     serializer = CommentSerializer(data=request.data)
     if serializer.is_valid(raise_exception=True):
-        serializer.save(article=article)
+        serializer.save(article=article, user=request.user)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+
+@api_view(['GET'])
+def article_comments(request, article_pk):
+    comments = Comment.objects.filter(article_id=article_pk)
+    serializer = CommentSerializer(comments, many=True)
+    return Response(serializer.data)
